figures/conv2d/requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.3.0
scipy>=1.7.0
rich>=13.0.0 